Add weekend events and modularize ageUp
